VillainAI/clone superai/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests
pip install deepseek-ai
